api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymysql==1.1.0
SQLAlchemy==2.0.23
alembic==1.12.1
cryptography==41.0.7
netmiko==4.2.0
paramiko==3.3.1
APScheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
emails==0.6.0
aiofiles==23.2.1